Finish rewrite of logroller
Rewrote logroller to use bash, as I was stupid the whole time and realized the RIO does have bash as a shell.
Also added various messages mainly for debugging and sanity checking, as waiting with no indication to what the script was doing was getting old.
Finally, rewrote the postdeploy artifact so that it works (hopefully, still needs testing). I'll look into a way to skip the task if it fails so it doesn't fail the whole build.